+++ postgresql-flexible-server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Release v1.5.0
+
+## Provider & Terraform Upgrade
+- Azurerm provider: 3.108.0 (#15)
+- Terraform: 1.8.5 (#15)
+## Enhancements
+- Add private endpoint (#15)
+   
 ## Release v1.4.0
 
 ## Provider & Terraform Upgrade